refactor(lint): extract stateless leaf converters into NodeFactory
Move the pure OA-annotation → leaf-node converters (exampleNode, header, link,
and the parameter/schema example collectors) out of SpecTreeBuilder into a new
stateless NodeFactory, mirroring the existing SchemaAccessor helper in the same
namespace. SpecTreeBuilder keeps only the stateful, component-index-aware
traversal; the moved methods depend on nothing but their argument.

The two near-identical example collectors now share a private examplesFromList
helper. SpecTreeBuilder drops ~117 lines (1166 → 1049). No behaviour change.
